+++ src/screen_title.h
@@ -13,11 +13,11 @@ public:
 	const char *getName() const override { return "TitleScreen"; }
 
 private:
-	static constexpr i32 BANNER_STARTPOS = -150.0f;
-	static constexpr i32 BANNER_ENDPOS   = 32.0f;
+	static constexpr i32 BANNER_STARTPOS = -150;
+	static constexpr i32 BANNER_ENDPOS   = 32;
 
 	GameMode mode_{ GameMode::LOREM };
 	u64 hiscore_;
-	f32 bannerPosition_ = BANNER_STARTPOS;
+	i32 bannerPosition_ = BANNER_STARTPOS;
 	f32 animTime_       = 0.0f;
 };
